chore: accountManager 기능별 분리
