+++ KnockKnock-iOS/Repository/SNS/AppleLoginManager.swift
@@ -1,0 +1,14 @@
+//
+//  AppleLoginManager.swift
+//  KnockKnock-iOS
+//
+//  Created by Daye on 2022/12/23.
+//
+
+import Foundation
+
+import AuthenticationServices
+
+protocol AppleLoginManagerProtocol {
+
+}
